% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poismf.R
\name{predict_all}
\alias{predict_all}
\title{Predict whole input matrix}
\usage{
predict_all(model)
}
\arguments{
\item{model}{A Poisson factorization model as output byfunction `poismf`.}
}
\value{
A matrix A x B with the full predictions for all rows and column. If the inputs
did not have numbers as IDs, the equivalences to their IDs in the outputs are in the
`poismf` object under fields `levels_A` and `levels_B`.
}
\description{
Outputs the predictions for the whole input matrix to which the model was fit.
Note that this will be a dense matrix, and in typical recommender systems scenarios will
likely not fit in memory.
}
