% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poismf.R
\name{print.poismf}
\alias{print.poismf}
\title{Get information about poismf object}
\usage{
\method{print}{poismf}(x, ...)
}
\arguments{
\item{x}{An object of class "poismf" as returned by function "poismf".}

\item{...}{Extra arguments (not used).}
}
\description{
Print basic properties of a "poismf" object.
}
