% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poismf.R
\name{summary.poismf}
\alias{summary.poismf}
\title{Get information about poismf object}
\usage{
\method{summary}{poismf}(object, ...)
}
\arguments{
\item{object}{An object of class "poismf" as returned by function "poismf".}

\item{...}{Extra arguments (not used).}
}
\description{
Print basic properties of a "poismf" object (same as `print.poismf` function).
}
\seealso{
\link{print.poismf}
}
